bot.py

import os
from discord.ext import commands
from dotenv import load_dotenv

load_dotenv()
TOKEN = os.getenv("BOT_TOKEN")

intents = commands.Intents.all()
bot = commands.Bot(command_prefix='!', intents=intents)

@bot.event
async def on_ready():
    print(f'Veramon Reunited is online as {bot.user}')
    print('Created by Killerdash117')

bot.run(TOKEN)
