requirements.txt
discord.py
python-dotenv
SQLAlchemy
